+++ katana/server.py
@@ -12,10 +12,12 @@ file that was distributed with this source code.
 from __future__ import absolute_import
 
 import logging
+import os
 
 from collections import namedtuple
 from multiprocessing import cpu_count
 
+import gevent
 import zmq.green
 
 from gevent.threadpool import ThreadPool
@@ -216,14 +218,12 @@ class ComponentServer(object):
 
         return payload
 
-    def __process_request(self, stream):
+    def __process_request_stream(self, stream):
         try:
             frames = Frames(*stream)
         except:
             LOG.error('Received an invalid multipart stream')
-            response = create_error_response('Failed to handle request')
-            self._send_response(response)
-            return
+            return create_error_response('Failed to handle request')
 
         # Update global schema registry when mappings are sent
         if frames.mappings:
@@ -233,28 +233,43 @@ class ComponentServer(object):
         action = frames.action.decode('utf8')
         if action not in self.callbacks:
             # Return an error when action doesn't exist
-            response = create_error_response(
+            return create_error_response(
                 'Invalid action for component {}: "{}"',
                 self.component_title,
                 action,
                 )
-            self._send_response(response)
-            return
 
         # Get command payload from request stream
         try:
             payload = unpack(frames.stream)
         except:
             LOG.exception('Received an invalid message format')
-            response = create_error_response('Internal communication failed')
-            self._send_response(response)
-            return
+            return create_error_response('Internal communication failed')
 
         payload = self.__process_request_payload(action, payload)
+        return [self.get_response_meta(payload) or EMPTY_META, pack(payload)]
 
-        self._send_response([
-            self.get_response_meta(payload) or EMPTY_META, pack(payload)
-            ])
+    def __process_request(self, stream, pid, timeout):
+        # Process request and get response stream.
+        # Request are processed inside a thread pool to avoid
+        # userland code to block requests.
+        res = self._pool.spawn(self.__process_request_stream, stream)
+
+        # Wait for a period of seconds to get the execution result
+        try:
+            response = res.get(timeout=timeout)
+        except gevent.Timeout:
+            msg = 'SDK execution timed out after {}ms'.format(
+                int(timeout * 1000),
+                pid,
+                )
+            response = create_error_response(msg)
+            LOG.warn('{}. PID: {}'.format(msg, pid))
+        except:
+            LOG.exception('Failed to handle request. PID: %d', pid)
+            response = create_error_response('Failed to handle request')
+
+        self._send_response(response)
 
     def process_payload(self, action, payload):
         """Process a request payload.
@@ -351,6 +366,9 @@ class ComponentServer(object):
 
         """
 
+        pid = os.getpid()
+        timeout = self.__args["timeout"] / 1000.0
+
         self.context = zmq.green.Context()
         self.poller = zmq.green.Poller()
 
@@ -370,10 +388,7 @@ class ComponentServer(object):
                 if events.get(self.__socket) == zmq.POLLIN:
                     # Get request multipart stream
                     stream = self.__socket.recv_multipart()
-                    # Process request and get response stream.
-                    # Request are processed inside a thread pool to avoid
-                    # userland code to block requests.
-                    self._pool.spawn(self.__process_request, stream)
+                    gevent.spawn(self.__process_request, stream, pid, timeout)
 
                 if events.get(self.__worker_socket) == zmq.POLLIN:
                     stream = self.__worker_socket.recv_multipart()